+++ test/integration/src/components/common/SelectBoxField/SelectBoxField.js
@@ -13,6 +13,13 @@ export class SelectBoxInput extends React.Component {
     this.props.handleChange(this.props.options[0].value);
   }
 
+  componentWillReceiveProps(nextProps) {
+    // First option always automatically picked in a select box.
+    if (nextProps.value == null) {
+      this.props.handleChange(this.props.options[0].value);
+    }
+  }
+
   render() {
     var i, value, options = [];
     for (i = 0; i < this.props.options.length; i++) {
@@ -20,11 +27,12 @@ export class SelectBoxInput extends React.Component {
       options.push(<option key={value} value={value}>{this.props.options[i].label}</option>);
     }
 
-    return <div>
+    return <div className={style.selectBox}>
       <label>{this.props.label}</label>
       <select value={this.props.value} onChange={(event) => this.props.handleChange(event.target.value, event)}>
         {options}
       </select>
+      {this.props.errorMessages.length > 0 ? <span>{this.props.errorMessages[0]}</span> : null}
     </div>;
   }
 }