+++ test/integration/src/components/common/TextField/NameField.js
@@ -2,15 +2,15 @@ import React from 'react';
 import createField from 'soya/lib/data/redux/form/createField';
 
 import { TextInput } from './TextField.js';
-import { isName } from '../../../validator/string.js';
+import { name } from '../../../validator/string.js';
 
 export default class NameInput extends React.Component {
   static connectId() {
     return 'NameField';
   }
 
-  static getOnChangeValidators() {
-    return [isName];
+  static getChangeValidators() {
+    return [name];
   }
 
   render() {